+++ NEWS.md
@@ -7,6 +7,13 @@ Master (not on release branches yet)
   (cs_f_user_boundary_conditions, cs_f_user_extra_operations,
   cs_user_f_initialization, usppmo, usipsu, and usipes).
 
+### Numerics:
+
+- Remove `idilat` 4 and 5 weakly compressible algorithm options.
+  * The validity domain of these options using a semi-analytic scheme
+    was too limited for future uses of the Burke-Schuman model for
+    fire modeling (gas combustion), so `idilat` 2 should be used now.
+
 Release 9.0.0 (unreleased)
 --------------------------